refactor: Spinner 데이터바인딩 적용
