+++ app/src/main/java/woowacourse/movie/home/view/MovieMainActivity.kt
@@ -1,12 +1,12 @@
-package woowacourse.movie
+package woowacourse.movie.home.view
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
+import woowacourse.movie.R
 import woowacourse.movie.databinding.ActivityMovieMainBinding
-import woowacourse.movie.home.view.MovieHomeFragment
 import woowacourse.movie.reservationlist.view.ReservationListFragment
 import woowacourse.movie.setting.view.SettingFragment
 
@@ -28,14 +28,17 @@ class MovieMainActivity : AppCompatActivity() {
                     replaceFragment(ReservationListFragment())
                     true
                 }
+
                 R.id.home_item -> {
                     replaceFragment(MovieHomeFragment())
                     true
                 }
+
                 R.id.setting_item -> {
                     replaceFragment(SettingFragment())
                     true
                 }
+
                 else -> false
             }
         }
